services/python-transaction/requirements.txt
alembic==1.3.1
argh==0.26.2
grpcio==1.25.0
grpcio-tools==1.25.0
Mako==1.1.0
MarkupSafe==1.1.1
pathtools==0.1.2
protobuf==3.10.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.1.2
six==1.13.0
SQLAlchemy==1.3.11
watchdog==0.9.0
xmlrunner==1.7.7
